pyproject.toml
[project]
name = "kato"
version = "0.1.0"
description = "Core-lib based developer agent for YouTrack, OpenHands, and Bitbucket."
requires-python = ">=3.11"
dependencies = [
  "core-lib>=0.2.0",
  "email-core-lib",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "jinja2>=3.1.6",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["kato*", "hydra_plugins*"]

[tool.setuptools.package-data]
"kato.templates.email" = ["*.j2"]
